Read and use CA Cert settings
The current release of cortexutils does not read nor use the setting for CA Certs from the Cortex UI. This PR is meant to solve this issue. It came about trying to use Cortex behind a MITM capable proxy. Sidenote,the UI should indicate that it's the path to the system CA bundle that should be added, not the whole CA Cert in Base64 format